+++ Missionframework/Interface/Permissions/Build_Permissions/cfgFunctions.hpp
@@ -1,0 +1,11 @@
+class Permisions_Build_UI 
+{
+    file = "Interface\Permissions\Build_Permissions\functions";
+
+    class build_perms_createAllButton {};
+    class build_perms_createNoneButton {};
+    class build_perms_createActiveText {};
+    class build_perms_createMenuRsc {};
+    class build_perms_loadMenu {};
+    class build_perms_unloadMenu {};
+};