+++ Missionframework/Interface/Permissions/General_Permissions/cfgFunctions.hpp
@@ -1,0 +1,11 @@
+class Permisions_General_UI 
+{
+    file = "Interface\Permissions\General_Permissions\functions";
+
+    class general_perms_createAllButton {};
+    class general_perms_createActiveText {};
+    class general_perms_createMenuRsc {};
+    class general_perms_createNoneButton {};
+    class general_perms_loadMenu {};
+    class general_perms_unloadMenu {};
+};